refactor pytest fixtures for new tests running threads
