+++ sherpa/astro/tests/test_astro.py
@@ -673,113 +673,13 @@ def test_missmatch_arf(make_data_path):
     assert parvals[2] == approx(2.35452, rel=1.0e-3)
 
 
-# for running regression tests from sherpa-test-data
-#
-# This is not a fixture, at least for now. The use of
-# SherpaTestCase.datadir is not ideal.
-#
-def run_thread(name, scriptname='fit.py'):
-    """Run a regression test from the sherpa-test-data submodule.
-
-    Parameters
-    ----------
-    name : string
-       The name of the science thread to run (e.g., pha_read,
-       radpro). The name should match the corresponding thread
-       name in the sherpa-test-data submodule. See examples below.
-    scriptname : string
-       The suffix of the test script file name, usually "fit.py."
-
-    Returns
-    -------
-    localsyms : dict
-        Any model parameters created by the script.
-
-    Examples
-    --------
-    Regression test script file names have the structure
-    "name-scriptname.py." By default, scriptname is set to "fit.py."
-    For example, if one wants to run the regression test
-    "pha_read-fit.py," they would write
-
-    >>> run_thread("pha_read")
-
-    If the regression test name is "lev3fft-bar.py," they would do
-
-    >>> run_thread("lev3fft", scriptname="bar.py")
-
-    """
-
-    basedir = SherpaTestCase.datadir
-    if basedir is None:
-        raise RuntimeError("Test needs the requires_data decorator")
-
-    scriptname = name + "-" + scriptname
-    cwd = os.getcwd()
-    os.chdir(basedir)
-
-    # Need to add to localsyms so that the scripts can work, but we
-    # do not need (for now) to return all the local symbols, so
-    # also have a version just for model parameters.
-    #
-    localsyms = {}
-    modelsyms = {}
-
-    def assign_model(name, val):
-        localsyms[name] = val
-        modelsyms[name] = val
-
-    old_assign_model = ui.get_model_autoassign_func()
-
-    try:
-        with open(scriptname, "rb") as fh:
-            cts = fh.read()
-        ui.set_model_autoassign_func(assign_model)
-        exec(compile(cts, scriptname, 'exec'), {}, localsyms)
-    finally:
-        ui.set_model_autoassign_func(old_assign_model)
-        os.chdir(cwd)
-
-    return modelsyms
-
-
-@pytest.fixture
-def clean_astro_ui():
-    """Ensure sherpa.astro.ui.clean is called before AND after the test.
-
-    This also resets the XSPEC settings (if XSPEC support is provided).
-
-    Notes
-    -----
-    It does NOT change the logging level; perhaps it should, but the
-    screen output is useful for debugging at this time.
-    """
-
-    # old_lgr_level = logger.getEffectiveLevel()
-    # logger.setLevel(logging.CRITICAL)
-
-    if has_xspec:
-        old_xspec = xspec.get_xsstate()
-    else:
-        old_xspec = None
-
-    ui.clean()
-    yield
-
-    ui.clean()
-    if old_xspec is not None:
-        xspec.set_xsstate(old_xspec)
-
-    # logger.setLevel(old_lgr_level)
-
-
 # This was test_threads.test_pileup
 #
 @requires_data
 @requires_fits
 @requires_xspec
 @pytest.mark.usefixtures("clean_astro_ui")
-def test_thread_pileup():
+def test_thread_pileup(run_thread):
 
     models = run_thread('pileup')
